client/requirements.txt
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
idna==2.10
pycodestyle==2.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
websocket-client==0.57.0
